* master: (42 commits)
  Updated CHANGELOG
  Fixes #273 by not copying the shared folder/file from the release if it has already been copied
  Re-added postgres env
  Testing only 1 env
  Test
  Test
  Changed cat command for travis
  Fix code climate issues
  Changed build image
  Fixes #275 by checking that the --no-suggest option exists before attempting to use it
  Fixes #276 - Allows the passphrase for the private key to be set
  Fixed missing reason of rollback (#270)
  Updated dependencies
  Updated dependencies
  Update README.md
  Update README.md
  Update README.md
  Update README.md
  Last chance
  Setting a random JWT
  ...

+++ resources/scripts/deploy/ShareFile.sh
@@ -1,6 +1,9 @@
 ### Created shared files - {{ deployment }}
 if [ -f {{ target_file }} ]; then
-    cp -pRn {{ target_file }} {{ source_file }}
+    if [ ! -f {{ source_file }} ]; then
+        cp -pRn {{ target_file }} {{ source_file }}
+    fi
+
     rm -rf {{ target_file }}
 fi
